feat: adaptação do problema das NRaininhas para genes do tipo double

+++ src/main/java/br/com/cefetmg/ag/Individuo.java
@@ -11,5 +11,5 @@ public interface Individuo {
 
     boolean isMaximizacao();
 
-    int[] getGenes();
+    double[] getGenes();
 }
